chore(api): move DTO to DTOs folder, fix PhotosController attributes, remove console project
