+++ PhotoFinder.Api/PhotoFinderAPI.Tests/PhotographerControllerTests.cs
@@ -1,4 +1,4 @@
-using PhotoFinderAPI.Controllers;
+using PhotoFinderAPI.DTOs;
 using PhotoFinderAPI.Models;
 using FluentAssertions;
 using Microsoft.AspNetCore.Mvc;
