+++ PhotoFinder.Api/PhotoFinderAPI.sln
@@ -4,8 +4,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoFinderAPI", "PhotoFind
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoFinderAPI.Tests", "PhotoFinderAPI.Tests\PhotoFinderAPI.Tests.csproj", "{2EA7CD61-15D5-4D5E-B8C6-FC64FD0BA8A1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PhotoFinderAPI.Console", "PhotoFinderAPI.Console\PhotoFinderAPI.Console.csproj", "{D96FF1A6-44CE-4B5A-B146-764ACC14132D}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,18 +38,6 @@ Global
 		{2EA7CD61-15D5-4D5E-B8C6-FC64FD0BA8A1}.Release|x64.Build.0 = Release|Any CPU
 		{2EA7CD61-15D5-4D5E-B8C6-FC64FD0BA8A1}.Release|x86.ActiveCfg = Release|Any CPU
 		{2EA7CD61-15D5-4D5E-B8C6-FC64FD0BA8A1}.Release|x86.Build.0 = Release|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Debug|x64.Build.0 = Debug|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Debug|x86.Build.0 = Debug|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Release|x64.ActiveCfg = Release|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Release|x64.Build.0 = Release|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Release|x86.ActiveCfg = Release|Any CPU
-		{D96FF1A6-44CE-4B5A-B146-764ACC14132D}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
