+++ PhotoFinder.Api/PhotoFinderAPI/Controllers/PhotosController.cs
@@ -1,0 +1,15 @@
+using Microsoft.AspNetCore.Mvc;
+using PhotoFinderAPI.Models;
+
+namespace PhotoFinderAPI.Controllers;
+
+[ApiController]
+[Route("api/[controller]")]
+public class PhotosController : ControllerBase
+{
+    [HttpGet]
+    public ActionResult<List<Photographer>> GetAll()
+    {
+            return Ok();
+    }
+}
